[#862] Simplify LambdaMatcher
