+++ src/test/java/org/mockito/internal/matchers/ArgumentMatcherPredicateTest.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (C) 2016 Mockito contributors.
+ *
+ * Licensed under the Apache License, Version 2.0.
+ */
+package org.mockito.internal.matchers;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.internal.matchers.domain.ShipSearchCriteria;
+import org.mockito.internal.matchers.domain.TacticalStation;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.Mockito.verify;
+
+public class ArgumentMatcherPredicateTest {
+
+    @Rule
+    public MockitoRule mockitoRule = MockitoJUnit.rule();
+
+    @Mock
+    private TacticalStation ts;
+
+    private ShipSearchCriteria searchCriteria = new ShipSearchCriteria(1000, 4);
+
+    @Test
+    public void shouldAllowToUseLambdaInMatcher() {
+        //when
+        int numberOfShips = ts.findNumberOfShipsInRangeByCriteria(searchCriteria);
+        //then
+        assertThat(numberOfShips).isZero();
+        //and
+        verify(ts).findNumberOfShipsInRangeByCriteria(argThat(c -> c.getMinimumRange() < 2000));
+    }
+
+    @Test
+    public void shouldAllowToUseLambdaInMatcherWithAdditionalDescription() {
+        //when
+        int numberOfShips = ts.findNumberOfShipsInRangeByCriteria(searchCriteria);
+        //then
+        assertThat(numberOfShips).isZero();
+        //and
+        verify(ts).findNumberOfShipsInRangeByCriteria(argThat(c -> c.getMinimumRange() < 2000, "minimum range closer than 2000"));
+    }
+
+    @Test
+    public void shouldAllowToUseLambdaWithMultipleConditionsInMatcher() {
+        //when
+        int numberOfShips = ts.findNumberOfShipsInRangeByCriteria(searchCriteria);
+        //then
+        assertThat(numberOfShips).isZero();
+        //and
+        verify(ts).findNumberOfShipsInRangeByCriteria(argThat(
+                c -> c.getMinimumRange() < 2000 && c.getNumberOfPhasers() > 2,
+                "ShipSearchCriteria minimumRange<2000 and numberOfPhasers>2"));
+    }
+
+}
