+++ src/test/java/org/mockito/internal/matchers/DescriptableWrapperMatcherTest.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (C) 2016 Mockito contributors.
+ *
+ * Licensed under the Apache License, Version 2.0.
+ */
+package org.mockito.internal.matchers;
+
+import org.assertj.core.api.ThrowableAssert;
+import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.internal.matchers.domain.ShipSearchCriteria;
+import org.mockito.internal.matchers.domain.TacticalStation;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.Mockito.verify;
+
+public class DescriptableWrapperMatcherTest {
+
+    @Rule
+    public MockitoRule mockitoRule = MockitoJUnit.rule();
+
+    @Mock
+    private TacticalStation ts;
+
+    private ShipSearchCriteria searchCriteria = new ShipSearchCriteria(1000, 4);
+
+    @Test
+    public void shouldKeepDescriptionInErrorMessage() {
+        //given
+        final String DESCRIPTION = "minimum range closer than 100";
+        //when
+        ts.findNumberOfShipsInRangeByCriteria(searchCriteria);
+        //then
+        ThrowableAssert.ThrowingCallable verifyLambda = () -> {
+            verify(ts).findNumberOfShipsInRangeByCriteria(argThat(c -> c.getMinimumRange() < 100, DESCRIPTION));
+        };
+        assertThatThrownBy(verifyLambda)
+                .isInstanceOf(AssertionError.class)
+                .hasMessageContaining("Argument(s) are different! Wanted:\n" +
+                        "ts.findNumberOfShipsInRangeByCriteria(\n" +
+                        "    " + DESCRIPTION + "\n" +
+                        ");")
+                .hasMessageContaining("Actual invocation has different arguments:\n" +
+                        "ts.findNumberOfShipsInRangeByCriteria(\n" +
+                        "    ShipSearchCriteria{minimumRange=1000, numberOfPhasers=4}\n" +
+                        ");");
+
+    }
+}
